fix (Person): RP fixes
