+++ Pds/Pds.Api.Contracts/Bill/BillContentDto.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Pds.Core.Enums;
 
 namespace Pds.Api.Contracts.Bill
 {
