+++ Pds/Pds.Api.Contracts/Content/ContentListBillDto.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
-using Pds.Core.Enums;
+﻿using Pds.Core.Enums;
 
 namespace Pds.Api.Contracts.Content
 {
