+++ Pds/Pds.Api.Contracts/Content/GetContentsResponse.cs
@@ -1,5 +1,4 @@
 ﻿using Pds.Api.Contracts.Paging;
-using Pds.Api.Contracts.Person;
 
 namespace Pds.Api.Contracts.Content
 {
