+++ Pds/Pds.Api.Contracts/Cost/GetCostsResponse.cs
@@ -1,5 +1,4 @@
 ﻿using Pds.Api.Contracts.Paging;
-using Pds.Api.Contracts.Person;
 
 namespace Pds.Api.Contracts.Cost
 {
