+++ Pds/Pds.Api.Contracts/Person/ResourceDto.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 
 namespace Pds.Api.Contracts.Person
 {
