+++ Pds/Pds.Api.Contracts/Person/UpdatePersonRequest.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 
 namespace Pds.Api.Contracts.Person
 {
