+++ Pds/Pds.Api/AppStart/SwaggerExtensions.cs
@@ -2,13 +2,9 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.OpenApi.Models;
-using Pds.Data;
 
 namespace Pds.Api.AppStart
 {
