+++ Pds/Pds.Api/Controllers/ApiControllerBase.cs
@@ -2,8 +2,6 @@
 using System.Net;
 using Pds.Core.Exceptions;
 using Microsoft.AspNetCore.Mvc;
-using Pds.Api.Authentication;
-using Pds.Core.Exceptions.Person;
 
 namespace Pds.Api.Controllers
 {
