+++ Pds/Pds.Api/appsettings.json
@@ -15,6 +15,6 @@
     "AllowedAppId": "tzfWOiWpwcbm05QhREjkyFQet3KzOyHl"
   },
   "ConnectionStrings": {
-    "DefaultConnection": "Server=localhost;Database=Pds;Trusted_Connection=True;"
+    "DefaultConnection": "Server=localhost;Initial Catalog=pds;User ID=admin;Password=123qweQWE;MultipleActiveResultSets=False;"
   }
 }
