+++ Pds/Pds.Data/QueryCreators/Person/PersonsFieldName.cs
@@ -1,4 +1,4 @@
-﻿namespace Pds.Data.Repositories
+﻿namespace Pds.Data.QueryCreators.Settings
 {
     public enum PersonsFieldName
     {
