+++ Pds/Pds.Data/Repositories/BrandRepository.cs
@@ -1,8 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
-using Pds.Data.Entities;
+﻿using Pds.Data.Entities;
 using Pds.Data.Repositories.Interfaces;
 
 namespace Pds.Data.Repositories
