+++ Pds/Pds.Data/Repositories/ClientRepository.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
-using Pds.Core.Extensions;
 using Pds.Data.Entities;
 using Pds.Data.Repositories.Interfaces;
