+++ Pds/Pds.Data/Repositories/Interfaces/IBrandRepository.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-using Pds.Data.Entities;
+﻿using Pds.Data.Entities;
 
 namespace Pds.Data.Repositories.Interfaces
 {
