+++ Pds/Pds.Services/Services/ContentService.cs
@@ -5,7 +5,6 @@ using Pds.Core.Enums;
 using Pds.Data;
 using Pds.Data.Entities;
 using Pds.Services.Interfaces;
-using Pds.Services.Models;
 using Pds.Services.Models.Content;
 
 namespace Pds.Services.Services
