+++ Pds/Pds.Web/Pages/Persons/Create.razor
@@ -77,7 +77,7 @@ else
                         <InputText class="form-control mb-2" id="city" @bind-Value="resource.Url" required="required" placeholder="Ссылка" />
                     </div>
                     <div class="col-auto">
-                        <button @onclick="() => RemoveResource(resource)" class="btn btn-danger btn-sm mb-2">Удалить ссылку</button>
+                        <div @onclick="() => RemoveResource(resource)" class="btn btn-danger btn-sm mb-2">Удалить ссылку</div>
                     </div>
                 </div>
             }
