+++ Pds/Pds.Web/Pages/Persons/List/Body.razor
@@ -3,9 +3,7 @@
 @using Microsoft.AspNetCore.Components.WebAssembly.Authentication
 @using Pds.Api.Contracts.Person
 @using Pds.Web.Common
-@using Pds.Web.Pages.Persons.List
 @using Api.Contracts.Paging
-@using Pds.Web.Components
 @using static Pds.Web.Common.TitleExtension
 
 @inject IApiClient ApiClient
